# docs/swagger.yaml
definitions:
  github_com_michaelwp_student_attendance_internal_models.AbsentRequest:
    properties:
      class_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      reason:
        type: string
      request_date:
        type: string
      status:
        $ref: '#/definitions/github_com_michaelwp_student_attendance_internal_models.AbsentRequestStatus'
      student_id:
        type: string
      updated_at:
        type: string
    type: object
  github_com_michaelwp_student_attendance_internal_models.AbsentRequestStatus:
    enum:
    - pending
    - approved
    - rejected
    type: string
    x-enum-varnames:
    - AbsentRequestStatusPending
    - AbsentRequestStatusApproved
    - AbsentRequestStatusRejected
...
info:
  contact: {}
paths:
  /absent-requests:
    post:
      consumes:
      - application/json
      description: Create a new absence request
      parameters:
      - description: Absent request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_michaelwp_student_attendance_internal_models.AbsentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Absent request created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create absent request
      tags:
      - Absent Requests
  /absent-requests/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an absent request
      parameters:
      - description: Absent request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Absent request deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid absent request ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete absent request
      tags:
      - Absent Requests
    get:
      consumes:
      - application/json
      description: Retrieve a specific absent request by ID
      parameters:
      - description: Absent request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Absent request retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid absent request ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Absent request not found
          schema:
            additionalProperties: true
            type: object
      summary: Get absent request by ID
      tags:
      - Absent Requests
  /absent-requests/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of an absent request (approve/reject)
      parameters:
      - description: Absent request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status update with status field
        in: body
        name: status
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Absent request status updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or status value
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Absent request not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update absent request status
      tags:
      - Absent Requests
  /absent-requests/class-id/{classId}:
    get:
      consumes:
      - application/json
      description: Retrieve absent requests for a specific class
      parameters:
      - description: Class ID
        in: path
        name: classId
        required: true
        type: integer
      - default: 10
        description: Number of requests to return (max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of requests to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Absent requests retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid class ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get absent requests by class
      tags:
      - Absent Requests
  /absent-requests/pending:
    get:
      consumes:
      - application/json
      description: Retrieve all pending absent requests
      parameters:
      - default: 10
        description: Number of requests to return (max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of requests to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Pending absent requests retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get pending absent requests
      tags:
      - Absent Requests
  /absent-requests/student-id/{studentId}:
    get:
      consumes:
      - application/json
      description: Retrieve absent requests for a specific student
      parameters:
      - description: Student ID
        in: path
        name: studentId
        required: true
        type: string
      - default: 10
        description: Number of requests to return (max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of requests to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Absent requests retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Student ID is required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get absent requests by student
      tags:
      - Absent Requests
...
swagger: "2.0"